dnsdist: Reduce the dependency on dnsdist::configuration.hh
Getting rid of the ugly DNSQuestion stubs in the meantime.
